requirements.txt
openai
requests
pymysql
flaskapp
flask
Flask==1.1.2
click==7.1.2
twilio
flask_cors
http.client
click==7.1.2
itsdangerous==2.0.1
jinja2==2.11.3
markupSafe==1.1.1
Werkzeug==1.0.1
python-dotenv>=0.19.0
flask-cors